extras: Stop running uWSGI as the ideascube user
Running all of uWSGI under our own user is pretty terrible, because of
the potential conflicts with other apps deployed on the same system.

It also creates problems when the ideascube and uwsgi packages are
installed as part of the same dpkg transaction.

As it turns out, we don't need to run all of uWSGI as ideascube: running
our app as the ideascube user is enough!

And in fact, our app was already configured to run under our own user,
so all we need to do is remove the uwsgi.service.d snippet, and tell
uWSGI to use its PAM plugin.

With that done, we don't need to chown /var/log/uwsgi to the ideascube
user any more.

Fixes #296