debian: Silence some more lintian problems
We're a web app. As most web app developers do, we bundle our static
assets: fonts and javascript dependencies.

Of course this is frowned upon by the Debian policy.

But we're ok with that, so let's just tell lintian to ignore these
issues, so that we can spot the real problems in its output.