debian: Silence some lintian problems
We build our package with dh-virtualenv, bundling all our dependencies
in a big Python virtual environment in /usr/share/python/ideascube/ .

As a result, we are bound to have some arch-dependent files in there,
namely the compiled CPython extensions.

For the same reason the shebangs for scripts in the virtual environment
will all point to the python executable in that virtual environment,
which is not the standard path.

Finally, we include the license files for all our dependencies, right
there in the virtual environment.

There's nothing we can do about it, unless we stop building with
dh-virtualenv.

But as long as we do, let's tell lintian to ignore these issues, so that
we can spot the real problems in its output.